+++ cartography/intel/aws/ec2/vpc.py
@@ -1,0 +1,159 @@
+import logging
+
+from .util import _get_botocore_config
+from cartography.util import run_cleanup_job
+
+logger = logging.getLogger(__name__)
+
+
+def get_from_aws(boto3_session, region):
+    client = boto3_session.client('ec2', region_name=region, config=_get_botocore_config())
+    # paginator not supported by boto
+    return client.describe_vpcs()
+
+
+def update_graph(neo4j_session, data, region, current_aws_account_id, aws_update_tag):
+    # https://docs.aws.amazon.com/cli/latest/reference/ec2/describe-vpcs.html
+    # {
+    #     "Vpcs": [
+    #         {
+    #             "VpcId": "vpc-a01106c2",
+    #             "InstanceTenancy": "default",
+    #             "Tags": [
+    #                 {
+    #                     "Value": "MyVPC",
+    #                     "Key": "Name"
+    #                 }
+    #             ],
+    #             "CidrBlockAssociations": [
+    #                 {
+    #                     "AssociationId": "vpc-cidr-assoc-a26a41ca",
+    #                     "CidrBlock": "10.0.0.0/16",
+    #                     "CidrBlockState": {
+    #                         "State": "associated"
+    #                     }
+    #                 }
+    #             ],
+    #             "State": "available",
+    #             "DhcpOptionsId": "dopt-7a8b9c2d",
+    #             "CidrBlock": "10.0.0.0/16",
+    #             "IsDefault": false
+    #         }
+    #     ]
+    # }
+
+    ingest_vpc = """
+    MERGE (new_vpc:AWSVpc{id: {VpcId}})
+    ON CREATE SET new_vpc.firstseen = timestamp(), new_vpc.vpcid ={VpcId}
+    SET new_vpc.instance_tenancy = {InstanceTenancy},
+    new_vpc.state = {State},
+    new_vpc.is_default = {IsDefault},
+    new_vpc.primary_cidr_block = {PrimaryCIDRBlock},
+    new_vpc.dhcp_options_id = {DhcpOptionsId},
+    new_vpc.region = {Region},
+    new_vpc.lastupdated = {aws_update_tag}
+    WITH new_vpc
+    MATCH (awsAccount:AWSAccount{id: {AWS_ACCOUNT_ID}})
+    MERGE (awsAccount)-[r:RESOURCE]->(new_vpc)
+    ON CREATE SET r.firstseen = timestamp()
+    SET r.lastupdated = {aws_update_tag}"""
+
+    for vpc in data['Vpcs']:
+        vpc_id = vpc["VpcId"]  # fail if not present
+
+        neo4j_session.run(
+            ingest_vpc,
+            VpcId=vpc_id,
+            InstanceTenancy=vpc.get("InstanceTenancy", None),
+            State=vpc.get("State", None),
+            IsDefault=vpc.get("IsDefault", None),
+            PrimaryCIDRBlock=vpc.get("CidrBlock", None),
+            DhcpOptionsId=vpc.get("DhcpOptionsId", None),
+            Region=region,
+            AWS_ACCOUNT_ID=current_aws_account_id,
+            aws_update_tag=aws_update_tag,
+        )
+
+        load_cidr_association_set(
+            neo4j_session,
+            vpc_id=vpc_id,
+            block_type="ipv4",
+            vpc_data=vpc,
+            aws_update_tag=aws_update_tag,
+        )
+
+        load_cidr_association_set(
+            neo4j_session,
+            vpc_id=vpc_id,
+            block_type="ipv6",
+            vpc_data=vpc,
+            aws_update_tag=aws_update_tag,
+        )
+
+
+def load_cidr_association_set(neo4j_session, vpc_id, vpc_data, block_type, aws_update_tag):
+    ingest_statement = _get_cidr_association_statement(block_type)
+
+    if block_type == "ipv6":
+        data = vpc_data.get("Ipv6CidrBlockAssociationSet", [])
+    else:
+        data = vpc_data.get("CidrBlockAssociationSet", [])
+
+    neo4j_session.run(
+        ingest_statement,
+        VpcId=vpc_id,
+        CidrBlock=data,
+        aws_update_tag=aws_update_tag,
+    )
+
+
+def _get_cidr_association_statement(block_type):
+    ingest_cidr = """
+    MATCH (vpc:AWSVpc{id: {VpcId}})
+    WITH vpc
+    UNWIND {CidrBlock} as block_data
+        MERGE (new_block:#BLOCK_TYPE#{id: {VpcId} + '|' + block_data.#BLOCK_CIDR#})
+        ON CREATE SET new_block.firstseen = timestamp()
+        SET new_block.association_id = block_data.AssociationId,
+        new_block.cidr_block = block_data.#BLOCK_CIDR#,
+        new_block.block_state = block_data.#STATE_NAME#.State,
+        new_block.block_state_message = block_data.#STATE_NAME#.StatusMessage,
+        new_block.lastupdated = {aws_update_tag}
+        WITH vpc, new_block
+        MERGE (vpc)-[r:BLOCK_ASSOCIATION]->(new_block)
+        ON CREATE SET r.firstseen = timestamp()
+        SET r.lastupdated = {aws_update_tag}"""
+
+    BLOCK_CIDR = "CidrBlock"
+    STATE_NAME = "CidrBlockState"
+
+    # base label type. We add the AWS ipv4 or 6 depending on block type
+    BLOCK_TYPE = "AWSCidrBlock"
+
+    if block_type == "ipv6":
+        BLOCK_CIDR = "Ipv6" + BLOCK_CIDR
+        STATE_NAME = "Ipv6" + STATE_NAME
+        BLOCK_TYPE = BLOCK_TYPE + ":AWSIpv6CidrBlock"
+    elif block_type == "ipv4":
+        BLOCK_TYPE = BLOCK_TYPE + ":AWSIpv4CidrBlock"
+    else:
+        raise ValueError(f"Unsupported block type specified - {block_type}")
+
+    return ingest_cidr.replace("#BLOCK_CIDR#", BLOCK_CIDR) \
+        .replace("#STATE_NAME#", STATE_NAME) \
+        .replace("#BLOCK_TYPE#", BLOCK_TYPE)
+
+
+def cleanup_graph(neo4j_session, common_job_parameters):
+    run_cleanup_job('aws_import_vpc_cleanup.json', neo4j_session, common_job_parameters)
+
+
+def sync(
+        neo4j_session, boto3_session, regions, current_aws_account_id, aws_update_tag,
+        common_job_parameters,
+):
+    for region in regions:
+        logger.debug("Syncing EC2 VPC for region '%s' in account '%s'.", region, current_aws_account_id)
+        data = get_from_aws(boto3_session, region)
+        update_graph(neo4j_session, data, region, current_aws_account_id, aws_update_tag)
+    cleanup_graph(neo4j_session, common_job_parameters)
